+++ ICSharpCode.AvalonEdit/CodeCompletion/CompletionWindowBase.cs
@@ -331,6 +331,9 @@ namespace ICSharpCode.AvalonEdit.CodeCompletion
 		protected void UpdatePosition()
 		{
 			TextView textView = this.TextArea.TextView;
+			if (PresentationSource.FromVisual(textView) == null)
+				return;
+			
 			// PointToScreen returns device dependent units (physical pixels)
 			Point location = textView.PointToScreen(visualLocation - textView.ScrollOffset);
 			Point locationTop = textView.PointToScreen(visualLocationTop - textView.ScrollOffset);